assignment-03/CUS/pyproject.toml
[project]
name = "control-unit-subsystem"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.2",
    "paho-mqtt>=2.1.0",
    "pyserial-asyncio>=0.6",
    "uvicorn[standard]>=0.38.0",
]
